requirements.txt
numpy==1.22.3
flatbuffers==2.0
eclipse-zenoh-nightly==0.6.0b120221124
Pillow==9.0.1